requirements.txt
streamlit
google-generativeai
PyMuPDF
tiktoken
faiss-cpu
sentence-transformers
gtts
